ajustando para subir arquivos no s3 aws

+++ src/controllers/PostController.js
@@ -10,20 +10,22 @@ module.exports = {
     },
     async store(req, res) {
         const { author, place, description, hashtags } = req.body
-        const { filename: image  } = req.file
-        
-        const [name] = image.split('.')
-        const fileName = `${name}.jpg`
+        const { filename: image } = req.file
+        let { location: fileName } = req.file
 
-        await sharp(req.file.path)
-            .resize(500)
-            .jpeg({ quality: 100 })
-            .toFile(
-                path.resolve(req.file.destination, 'resized', fileName)
-            )
-
-        fs.unlinkSync(req.file.path)
-
+        if (image !== undefined) {
+            const [name] = image.split('.')
+            fileName = `${name}.jpg`
+    
+            await sharp(req.file.path)
+                .resize(500)
+                .jpeg({ quality: 100 })
+                .toFile(
+                    path.resolve(req.file.destination, 'resized', fileName)
+                )
+    
+            fs.unlinkSync(req.file.path)
+        }
         const post = await Post.create({
             author,
             place,
